pyproject.toml
[project]
name = "soothe"
version = "0.1.0"
description = "Multi-agent harness built on deepagents and langchain/langgraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
keywords = ["agents", "ai", "llm", "langgraph", "langchain", "multi-agent", "subagents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "deepagents>=0.4.10,<1.0.0",
    "langchain-core>=1.2.18,<2.0.0",
    "langchain>=1.2.11,<2.0.0",
    "langchain-openai>=0.3.0,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "langgraph>=1.1.1,<2.0.0",
    "langchain-mcp-adapters>=0.2.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "langchain-tavily>=0.2.17",
    "typer[all]>=0.9.0,<1.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "textual>=0.40.0",
]

[project.optional-dependencies]
research = [
    "tavily-python>=0.5.0",
]
browser = [
    "browser-use<=0.12.0",
]
claude = [
    "claude-agent-sdk>=0.1.40",
]
serper = [
    "requests>=2.31.0",
]
jina = [
    "aiohttp>=3.9.0",
]
media = [
    "pillow>=10.0.0",
    "openai>=1.0.0",
]
rocksdb = [
    "rocksdict>=0.3",
]
pgvector = [
    "psycopg[pool]>=3.1",
    "pgvector>=0.3",
]
weaviate = [
    "weaviate-client>=4.0",
]
ollama = [
    "langchain-ollama>=0.3.0",
]
all = [
    "soothe[research,browser,claude,serper,jina,media,rocksdb,pgvector,weaviate,ollama]",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/soothe"]

[project.scripts]
soothe = "soothe.cli:app"

[tool.ruff]
line-length = 120
exclude = []

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "ISC001",
    "PERF203",
    "SLF001",
    "PLR0913",
    "PLC0414",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true
known-first-party = ["soothe"]

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-var-parameters = true

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN001", "ANN201", "ANN202", "ARG002", "D1", "PLC0415", "PLR2004", "PT018", "S101", "S311"]
"src/soothe/protocols/**" = ["TC003", "ANN401"]
"src/soothe/planning/**" = ["ANN401", "BLE001", "D107", "TRY300"]
"src/soothe/policy/**" = ["D107", "PLR0911"]
"src/soothe/durability/**" = ["ANN401"]
"src/soothe/remote/**" = ["ANN401", "ARG002", "PLC0415", "TC003"]
"src/soothe/config.py" = ["ANN401", "PLC0415", "TC002"]
"src/soothe/agent.py" = ["ARG001", "C901", "PLC0415", "PLR0911", "SIM108", "TC001", "TC002", "TC003"]
"src/soothe/vector_store/**" = ["ANN401", "ARG002", "B905", "BLE001", "D107", "PLC0415", "S608", "TC001"]
"src/soothe/persistence/**" = ["ANN401", "PLC0415", "SIM105"]
"src/soothe/context/vector_context.py" = ["BLE001"]
"src/soothe/memory_store/vector_memory.py" = ["BLE001"]
"src/soothe/cli/**" = ["ANN", "ASYNC240", "BLE001", "C901", "D107", "E501", "FBT002", "G201", "PIE810", "PLC0415", "PLR0912", "PLR0915", "PLR1714", "PLR2004", "PTH123", "RUF012", "S110", "S603", "S607", "SIM102", "SIM105", "TCH003", "TRY300"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
